sketch out of prposed changed for scenario_switcher
